# OpenAPI.yaml
openapi: 3.0.3
info:
  title: GMT Tech Challenge API
  version: 1.0.0
  description: API for managing video metadata and statistics
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Authenticate user and receive JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '403':
          description: Authentication failed
        '500':
          description: Unhandled exceptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Credentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required:
        - accessToken
      properties:
        accessToken:
          type: string
          description: JWT authentication token
...
    ErrorResponse:
      type: object
      description: Response object for errors with messages in the body
      properties:
        status:
          type: integer
          format: int32
          description: HTTP Status Code
        message:
          type: string
          description: Error Message
        timestamp:
          type: integer
          format: int64
      required:
        - status
        - message
        - timestamp
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []